Add files via upload
The inspiration and credit for the idea of this program comes from the Raspberry Pi foundation; under the heading of Installations, the process of writing an image to an sd card or USB device is given.  As I had only raspbian to work with, i had to use the built in dd command. After many repetitive image writings and so on, I decided to write this program to automate this process. It also provides for SFTP of image or zip OS files to be retrieved. The program can write image or compressed zip files to disk. It can unzip and write to the USB or sd card device, as stated in the raspberry pi Installation section. This program is provided under the Apache 2.0 license and is provided AS-IS without any express warranties or expectations.  I hope you enjoy the fruits of my labor. I am not an experienced python programmer, as this is my first project. In summary, the program can: SFTP a remote file, or skip directly to writing an image to a USB or sd card. It has provisions for checking the written file via truncate and diff; will flush the write cache after checking the file and unmount and eject the USB or sd card for safe removal.