Add files via upload
updated to release status. this is the most current file that has been tested. It uses command line arguments to determine what is being written. refer to user documentation inside the program listing for instructions. New -- this version can now write .iso filetypes making use of 3 types of files.   Image, Zip and ISO for the windows version of Raspbian Desktop.